fix(routing): 4 bugs structurels NOD/graph — alignement parité mkgmap r4924
Verrouillé via diff binaire NOD subfile contre mkgmap-r4924 sur petite tuile
(BDTOPO-007-004.mp, 477 routes) : boundary_len 3789B → 342B (= mkgmap),
hcb_len 918B → 126B (= mkgmap), header NOD byte-identique.

Reste à traiter (chantier NOD1 rewrite) : encodage RouteCenter / RouteNode
deltas / RouteArc var-bit / Table A/B — totalement divergent de mkgmap,
empêche le routing côté firmware (BaseCamp + Alpha 100).

Fix 1 — imgforge/src/img/nod.rs : class_boundaries init u32::MAX → i32::MAX
  Le writer émet des deltas calculés sur des `u32::MAX` initiaux. Le firmware
  Garmin (C signé int32) lit le delta 0xFFFFFFFF comme -1 et rejette le graph.
  mkgmap utilise Integer.MAX_VALUE (0x7FFFFFFF) — on s'aligne.

Fix 2 — garmin-routing-graph/src/lib.rs : FNV-1a masqué à 31 bits
  coord_to_node_id générait des u32 > 2³¹-1. mkgmap RoadHelper.NodeIndex
  parse en Integer.parseInt (NumberFormatException sur "3871640276"). Le
  firmware Garmin lit aussi node_id en int32 — bit haut = offset négatif.

Fix 3 — mpforge/src/pipeline/routing_graph.rs : dedup_by_key sur node_id
  Deux points consécutifs d'une polyline peuvent se quantizer sur la même
  cellule grid → même node_id → arc self-loop. mkgmap warne explicitement
  « consecutive identical nodes - routing will be broken » et refuse de
  router. Le firmware Garmin également (arc de longueur zéro).

Fix 4 — imgforge graph_builder + writer pre_compute_routing : transmet le
  flag boundary depuis NodEntry du .mp
  graph_builder.build_graph_with_junctions forçait is_boundary=true sur tous
  les nœuds → boundary section NOD3 11× trop grosse → BaseCamp snap-fail
  sur un mauvais nœud-frontière. On collecte les coords flagged boundary=1
  dans le .mp (mpforge::is_boundary_point) et on ne marque que celles-là.

+++ tools/imgforge/src/img/nod.rs
@@ -259,7 +259,11 @@ impl NodWriter {
             table_a_positions: Vec::new(),
             nod1_data: None,
             node_offsets: Vec::new(),
-            class_boundaries: [u32::MAX; 5], // mkgmap: Integer.MAX_VALUE
+            // mkgmap initializes with Integer.MAX_VALUE (Java signed int = 0x7FFFFFFF).
+            // Using u32::MAX (0xFFFFFFFF) here caused the written delta classBoundaries[1] − [0]
+            // to be 0xFFFFFFFF, which the Garmin firmware reads as int32 = -1 and rejects the
+            // routing graph. Match mkgmap exactly.
+            class_boundaries: [i32::MAX as u32; 5],
         }
     }
